Fixes dbus linking on non-Linux platforms.
Commit 6e1d6bd6f51b3a66de029792433dcbfc9fe4a048 added "LIBPATH='.'"
overrides to most build setups, though the purpose was unclear.  The
SConstruct script already sets up LIBPATH to include the current
directory, as well as the proper directory for the dbus library when
dbus support is included, so the effect of the overrides was to undo
the latter.  This typically doesn't hurt on Linux, where the dbus
library is installed in a "system" library area which is searched by
default, but on OSX and BSD platforms it results in link errors.
Removing the overrides fixes this.

TESTED:
Tested builds and ran regression tests on all OSX versions from 10.5
through 10.11, as well as Ubuntu 14.04.1, FreeBSD 10.1, OpenBSD 5.6,
and NetBSD 6.1.5.