Proposing Dropping QT support
Implementation injected QT objects into the library in a
not very compataible way.  Implementation Injected QTSocket and QTTime
classes for socket communication and time checks.  The QTSocket was
stored in the main c struct as a void* but as alternate to the socket_t.

Multiple C source files had to be compiled by the c++ compiler with USE_QT
because they included c++ syntax in the C code files.

The QT Support for cross-platform compiling and warning hunting and
general code cleanup has been causing a lot of pain for a while now.

No developers have stood up recently to help sort out the QT build issues.